+++ dens.hpp
@@ -9,7 +9,7 @@ namespace Dens {
    * Compute 1D-densities for given observables.
    * Runs in parallel on all available GPUs.
    */
-  std::vector<std::vector<float>>
+  std::vector<std::vector<double>>
   compute_densities_1d(std::vector<std::size_t> selected_cols
                      , const float* coords
                      , std::size_t n_rows
@@ -18,7 +18,7 @@ namespace Dens {
   /**
    * Compute 2D/3D-densities for given observables.
    */
-  std::tuple<std::vector<std::vector<float>>, std::vector<std::string>>
+  std::tuple<std::vector<std::vector<double>>, std::vector<std::string>>
   compute_densities_nd(std::vector<std::size_t> selected_cols
                      , const float* coords
                      , std::size_t n_rows
